ccnmtldjango/template/requirements.txt
Django==1.7.4
httplib2==0.9
restclient==0.11.0
feedparser==5.1.3
Markdown==2.5.1
smartypants==1.8.6
uuid==1.30
psycopg2==2.5.4
Pillow==2.7.0
nose==1.3.4
versiontools==1.9.1
statsd==3.0.1
pep8==1.5.7
pyflakes==0.8.1
mccabe==0.2.1
flake8==2.2.5
tornado==3.1.1
BeautifulSoup==3.2.1
cssselect==0.9.1
lxml==3.2.3
fuzzywuzzy==0.4.0
sure==1.2.7
ipython==1.2.1
ipdb==0.8
rdflib==4.1.2
selenium==2.44.0
coverage==3.7.1
requests==2.4.3
logilab-common==0.63.0
logilab-astng==0.24.3
astroid==1.2.1
pylint==1.3.1
six==1.8.0
factory_boy==2.4.1
python-ldap==2.4.18
sqlparse==0.1.13
boto==2.34.0
python-dateutil==2.2

djangowind==0.13.2
sorl==3.1
tagging==0.3-pre
typogrify==2.0.7
django-indexer==0.3.0
django-templatetag-sugar==1.0
django-paging==0.2.5
django-annoying==0.8.0
django-nose==1.2
django-appconf==0.6
django_compressor==1.4
django-statsd-mozilla==0.3.14
raven==5.1.1
django-bootstrap3==5.0.2
django-bootstrap-form==3.1.0
django-debug-toolbar==1.2.2
django-waffle==0.10
django-jenkins==0.16.3
django-smoketest==0.3.0
django-extensions==1.4.6
django-stagingcontext==0.1.0
django-ga-context==0.1.0
django-impersonate==0.9
django-registration==1.1
django-treebeard==2.0
django-pagetree==1.0.4
django-pageblocks==1.0.0
django-quizblock==1.0.1
django-markwhat==2014.9.20
gunicorn==19.1.1
django-infranil==0.1.0
django-flatblocks==0.9
django-storages==1.1.8
